stream_unsubscribe: fix last admin unsubscribing from private stream
Previously when a private stream had only one administrator, that
administrator could still unsubscribe, this caused the stream to
only heve users who dont have permissions to invite anyone to the
stream or remove anyone from it.

This adds a verification that the unsubscribing user is not the last
administrator in the stream, showing a notification and not
unsubscribing the user in case he is.

Fixes #11962